devtools/client/app-manager/test/test_device_store.html

<!DOCTYPE html>

<!--
Bug 901520 - [app manager] data store for device
-->

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8" src="chrome://devtools/content/app-manager/content/template.js"></script>
    <script type="application/javascript;version=1.8">
      const Cu = Components.utils;
      const {require} = Cu.import("resource://devtools/shared/Loader.jsm");
      const {DebuggerServer} = require("devtools/server/main");

      if (!DebuggerServer.initialized) {
        DebuggerServer.init();
        DebuggerServer.addBrowserActors();
      }

      function compare(o1, o2, msg) {
        is(JSON.stringify(o1), JSON.stringify(o2), msg);
      }

      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Cu.import("resource://gre/modules/Services.jsm");
        Cu.import("resource://devtools/client/framework/gDevTools.jsm");

        const {ConnectionManager} = require("devtools/shared/client/connection-manager");
        const DeviceStore = require("devtools/client/app-manager/device-store");

        let {getDeviceFront} = require("devtools/server/actors/device");

        let connection = ConnectionManager.createConnection();
        let store = new DeviceStore(connection);

        connection.once("connected", function() {
          store.on("set", function check(event, path, value) {
            if (path.join(".") != "description") return;
            store.off("set", check);
            info("Connected");
            connection.client.listTabs((resp) => {
              info("List tabs response");
              let deviceFront = getDeviceFront(connection.client, resp);
              deviceFront.getDescription().then(json => {
                info("getDescription response: " + JSON.stringify(json));
                json.dpi = Math.ceil(json.dpi);
                for (let key in json) {
                  compare(json[key], store.object.description[key], "description." + key + " is valid");
                  compare(json[key], value[key], "description." + key + " is valid");
                }
                connection.disconnect();
              }).then(null, (error) => ok(false, "Error:" + error));
            });
          });
        });

        connection.once("disconnected", function() {
          compare(store.object, {description:{},permissions:[],tabs:[]}, "empty store after disconnect")
          connection.destroy();
          DebuggerServer.destroy();
          SimpleTest.finish();
        });

        compare(store.object, {description:{},permissions:[],tabs:[]}, "empty store before disconnect")

        connection.connect();

      }

    </script>
  </body>
</html>
